requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==11.2.1
python-multipart==0.0.9
